backend/requirements.txt
aws-cdk-lib==2.177.0
constructs>=10.0.0,<11.0.0
aws-lambda-powertools==3.5.0
boto3==1.36.16
requests-aws4auth==1.3.1
opensearch-py==2.8.0
PyYAML==6.0.2
pytest==8.3.4
pytest-cov==4.1.0
moto==4.2.13  # For mocking AWS services
